+++ ChangeLog.md
@@ -8,6 +8,7 @@
 
     - Improve README
     - Define new 'assistant-checker' event, thus bumping minor candidate version number
+    - Provide the PCK object to the assistant-activated event
 
 ### 1.0.0